runtime-services/search-api/requirements.txt
fastapi
uvicorn[standard]
gunicorn
supabase
python-dotenv
pydantic
slowapi
